CI-71 Updated API to ELK 5

+++ src/Sfa.Das.ApprenticeshipInfoService.Infrastructure/Elasticsearch/ElasticsearchCustomClient.cs
@@ -47,7 +47,7 @@
                 Identifier = identifier,
                 ReturnCode = result.ApiCall?.HttpStatusCode,
                 Successful = result.ApiCall?.Success,
-                SearchTime = result.Took,
+                SearchTime = (int)result.Took,
                 RequestTime = Math.Round(clientRequestTime.TotalMilliseconds, 2),
                 MaxScore = result.MaxScore,
                 HitCount = result.Hits?.Count(),
